Change conversation template file structure
